refactor: consolidate streaming files into single module
Merge streaming_context.py, streaming_status.py, streaming_citations.py
into streaming.py. Reduces 4 files to 1 with all exports preserved.
Update all imports across runtime/agent, runtime/tools, and tests.
