+++ src/fleet_rlm/runtime/agent/forced_routing.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING, Any
 
 import dspy
 
-from fleet_rlm.runtime.execution.streaming_context import StreamingContext
+from fleet_rlm.runtime.execution.streaming import StreamingContext
 from fleet_rlm.runtime.models.streaming import StreamEvent
 
 from .chat_session_state import append_history, forced_delegate_context, history_turns
