+++ src/fleet_rlm/runtime/agent/recursive_runtime.py
@@ -23,8 +23,10 @@ from fleet_rlm.runtime.execution.streaming import (
     ReActStatusProvider,
     _process_stream_value,
 )
-from fleet_rlm.runtime.execution.streaming_citations import _normalize_trajectory
-from fleet_rlm.runtime.execution.streaming_context import StreamingContext
+from fleet_rlm.runtime.execution.streaming import (
+    StreamingContext,
+    _normalize_trajectory,
+)
 from fleet_rlm.runtime.models.streaming import StreamEvent
 
 from .delegation_policy import (
